api/pyproject.toml
[project]
name = "wump"
version = "0.1.0"
description = "wump: Who's Using My Package? - Dependency sponsorship discovery API"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "wump contributors" }
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
    # OpenTelemetry core and instrumentation
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-exporter-otlp>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
filterwarnings = [
    "ignore:coroutine 'Connection._cancel' was never awaited:RuntimeWarning",
]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
explicit_package_bases = true
